ai-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
openai==1.3.7
Pillow==10.1.0
PyMuPDF==1.23.8
reportlab==4.0.7
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
